Move opaque stub creation into its own file.
