+++ src/java_bytecode/java_bytecode_language.cpp
@@ -16,6 +16,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "java_bytecode_internal_additions.h"
 #include "java_bytecode_typecheck.h"
 #include "java_entry_point.h"
+#include "java_opaque_method_stubs.h"
 #include "java_bytecode_parser.h"
 #include "java_class_loader.h"
